feat: 실행부와 메인 로직 모듈 분리
